refactor(ActualButton): useSquareDetectionをclient/へ切り出す
ActualButton/ActualAnchorButton(将来分割予定)の双方から利用するため、
square検出ロジックをuseSquareDetection.tsとして独立させた。

返り値もonlyBodyからdataOnlyBodyAttr（onlyBody || undefined）にリネームし、
呼び出し側でdata-only-body属性に設定すべき値であることを名前から明確にした。
